refactor: redesign config interface
